+++ pyproject.toml
@@ -109,7 +109,13 @@ documentation = "https://github.com/bact/pitloom#readme"
 issues = "https://github.com/bact/pitloom/issues"
 
 [dependency-groups]
-lint = ["bandit>=1.9.4", "flake8>=7.3.0", "pylint>=4.0.7", "ruff>=0.16.3"]
+lint = [
+    "bandit>=1.9.4",
+    "flake8>=7.3.0",
+    "flake8-cognitive-complexity>=0.1.0",
+    "pylint>=4.0.7",
+    "ruff>=0.16.3",
+]
 test = [
     "check-wheel-contents>=0.6.3",
     "installer>=1.0.1",
@@ -230,13 +236,15 @@ init-hook = "import sys; sys.path.insert(0, 'src'); sys.path.insert(0, 'examples
 source-roots = ["src", "examples/sentimentdemo-aibom/src"]
 
 [tool.pylint.design]
-max-args = 8
+max-args = 6                  # Interim ceiling; AGENTS.md target is 5
 max-attributes = 12
-max-branches = 25
-max-locals = 20
-max-positional-arguments = 8
+max-branches = 20             # Interim ceiling; 2 functions still over even this -- see complexity-and-file-size-roadmap.md
+max-locals = 18               # Interim ceiling; AGENTS.md target is 15
+max-nested-blocks = 5         # Already at AGENTS.md's target -- pinned explicitly, not relying on pylint's default
+max-positional-arguments = 6  # Track max-args
 max-public-methods = 20
-max-statements = 100
+max-returns = 6                # Already at AGENTS.md's target -- pinned explicitly, not relying on pylint's default
+max-statements = 80           # Tightened from 100; zero current violations, room to tighten further later
 
 [tool.pylint.format]
 max-line-length = 88
@@ -312,17 +320,25 @@ ignore = [
     "E501", # line too long, handled by formatter
 ]
 select = [
-    "E",  # pycodestyle errors
-    "W",  # pycodestyle warnings
-    "F",  # pyflakes
-    "I",  # isort
-    "B",  # flake8-bugbear
-    "C4", # flake8-comprehensions
-    "UP", # pyupgrade
+    "E",   # pycodestyle errors
+    "W",   # pycodestyle warnings
+    "F",   # pyflakes
+    "I",   # isort
+    "B",   # flake8-bugbear
+    "C4",  # flake8-comprehensions
+    "UP",  # pyupgrade
+    "C90", # mccabe complexity -- see [tool.ruff.lint.mccabe] below
 ]
 
 [tool.ruff.lint.isort]
 known-first-party = ["pitloom"]
+
+[tool.ruff.lint.mccabe]
+# Interim ceiling, headroom above today's measured worst (32) so this
+# activates without breaking CI. AGENTS.md's real target is 10 -- see
+# working-docs/design/complexity-and-file-size-roadmap.md for the
+# 33-function backlog that has to shrink before this can drop.
+max-complexity = 35
 
 [tool.coverage.run]
 branch = true
